+++ CHANGELOG.md
@@ -8,7 +8,7 @@ To report an issue, or to see proposed and upcoming enhancements, check out [63K
 
 Report all vulnerabilities under the [Security menu](https://github.com/63Klabs/cache-data/security/advisories) in the Cache-Data GitHub repository.
 
-## v1.3.12 (unreleased)
+## v1.3.12 (2026-05-03)
 
 ### Fixed